add shared lib face_detect

+++ src/image_processing/image_processing.hpp
@@ -8,7 +8,7 @@
 
 #include <opencv2/opencv.hpp>
 #include <iostream>
-#include "../face_detect/face_detect.hpp"
+#include <library.h>
 
 
 namespace image_processing {///< @namespace image_processing
@@ -36,8 +36,8 @@ namespace image_processing {///< @namespace image_processing
         void image_processing(cv::Mat image, std::string pict_name);
         ~ImageProcessing() = default;
     private:
-        uint image_resize_cof_col; ///< @name image resize cof
-        uint image_resize_cof_row; ///< @name image resize cof
+        double image_resize_cof_col; ///< @name image resize cof
+        double image_resize_cof_row; ///< @name image resize cof
         /*!@brief resize picture and save it
          * @param[in] path path to file
          * @details function get path to file, verf is file can be open
